// src/store/interfaces.ts
type Info = {
  inputunit?: string;
  inputvalue: string;
  type?: string;
};

export type Attribute = {
  displayName: string;
  info: Info;
  mandatory?: string;
  name?: string;
  value?: string;
  writeAccess?: string;
};

export type Attributes = {
  [key: string]: Attribute;
};

export type RequestAttributeData = {
  [key: string]: string;
};

export type ConversionUnit = {
  label: string;
  multiplier: number;
  offset: number;
  type: string;
};

export type ConversionUnits = {
  [key: string]: ConversionUnit;
};

export type Units =
  typeof DIM_METER
  | typeof DIM_MM
  | typeof DIM_GRAM
  | typeof DIM_KILOGRAM
  | typeof DIM_MG
// ...
export type Validation = {
  statusOK: boolean;
  message: string;
};